Fixed some views issues

+++ API/Parking.Models/Models/Session.cs
@@ -109,6 +109,11 @@ namespace Sieena.Parking.API.Models
             using (EntityContext ctx = new EntityContext())
             {
                 Session sess = ctx.Sessions.Where(s => s.SessionId.Equals(sessionId)).FirstOrDefault();
+                if (sess == null)
+                {
+                    return null;
+                }
+
                 if (sess.ExpiresAt < DateTime.Now && sess.ExpiresAt != DateTime.MaxValue)
                 {
                     ctx.Sessions.DeleteObject(sess);
